+++ database/data.sql
@@ -5739,7 +5739,7 @@ CREATE TABLE `RelativeIceThickness` (
   KEY `RelativeIceThickness_fk_motionCorrectionId` (`motionCorrectionId`),
   CONSTRAINT `RelativeIceThickness_fk_motionCorrectionId` FOREIGN KEY (`motionCorrectionId`) REFERENCES `MotionCorrection` (`motionCorrectionId`) ON DELETE NO ACTION ON UPDATE CASCADE,
   CONSTRAINT `RelativeIceThickness_fk_programId` FOREIGN KEY (`autoProcProgramId`) REFERENCES `AutoProcProgram` (`autoProcProgramId`) ON DELETE NO ACTION ON UPDATE CASCADE
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
+) ENGINE=InnoDB AUTO_INCREMENT=4 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
 --
@@ -5748,6 +5748,7 @@ CREATE TABLE `RelativeIceThickness` (
 
 LOCK TABLES `RelativeIceThickness` WRITE;
 /*!40000 ALTER TABLE `RelativeIceThickness` DISABLE KEYS */;
+INSERT INTO `RelativeIceThickness` VALUES (1,23,56986680,1,2,3,4,5),(2,24,56986680,2,2,3,4,5),(3,25,56986680,3,4,5,6,7);
 /*!40000 ALTER TABLE `RelativeIceThickness` ENABLE KEYS */;
 UNLOCK TABLES;
 
@@ -8404,4 +8405,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2023-01-12  8:45:59
+-- Dump completed on 2023-01-19 11:21:41